# frontend/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "GitLab User Reports",
        "version": "0.1.0"
    },
    "paths": {
...
        "/projects/{project_id}/members": {
            "get": {
                "tags": ["projects"],
                "summary": "Get Project Members",
                "description": "Get members of a GitLab project.",
                "operationId": "get_project_members_projects__project_id__members_get",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "project_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "description": "GitLab project ID",
                            "title": "Project Id"
                        },
                        "description": "GitLab project ID"
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "minimum": 1,
                            "description": "Page number for pagination",
                            "default": 1,
                            "title": "Page"
                        },
                        "description": "Page number for pagination"
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "maximum": 100,
                            "minimum": 1,
                            "description": "Number of members per page",
                            "default": 20,
                            "title": "Per Page"
                        },
                        "description": "Number of members per page"
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "description": "Filter members by username or name",
                            "title": "Search"
                        },
                        "description": "Filter members by username or name"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/MembersResponse"
                                    },
                                    "title": "Response Get Project Members Projects  Project Id  Members Get"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid credentials",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponseModel"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Resource not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponseModel"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponseModel"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/projects/": {
            "get": {
                "tags": ["projects"],
                "summary": "List Projects",
                "description": "List GitLab projects with optional search filtering.",
                "operationId": "list_projects_projects__get",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "minimum": 1,
                            "description": "Page number for pagination",
                            "default": 1,
                            "title": "Page"
                        },
                        "description": "Page number for pagination"
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "maximum": 100,
                            "minimum": 1,
                            "description": "Number of projects per page",
                            "default": 20,
                            "title": "Per Page"
                        },
                        "description": "Number of projects per page"
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "description": "Filter projects by name or path",
                            "title": "Search"
                        },
                        "description": "Filter projects by name or path"
                    },
                    {
                        "name": "membership",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "type": "boolean"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "description": "Only return projects the authenticated user is a member of",
                            "title": "Membership"
                        },
                        "description": "Only return projects the authenticated user is a member of"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ProjectsResponse"
                                    },
                                    "title": "Response List Projects Projects  Get"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid credentials",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponseModel"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponseModel"
                                }
                            }
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponseModel"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/projects/{project_id}": {
            "get": {
                "tags": ["projects"],
                "summary": "Get Project",
                "description": "Get details of a specific GitLab project.",
                "operationId": "get_project_projects__project_id__get",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "project_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "description": "GitLab project ID",
                            "title": "Project Id"
                        },
                        "description": "GitLab project ID"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProjectsResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid credentials",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponseModel"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Resource not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponseModel"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponseModel"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/performance/projects/{project_id}": {
            "get": {
                "tags": ["performance"],
                "summary": "Get General Project Performance",
                "description": "Calculate and retrieve general project performance data over a specified time period.",
                "operationId": "get_general_project_performance_performance_projects__project_id__get",
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ],
                "parameters": [
                    {
                        "name": "project_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "description": "The ID of the GitLab project.",
                            "title": "Project Id"
                        },
                        "description": "The ID of the GitLab project."
                    },
                    {
                        "name": "start_date",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date-time",
                            "description": "The start date for the performance period in ISO format.",
                            "title": "Start Date"
                        },
                        "description": "The start date for the performance period in ISO format."
                    },
                    {
                        "name": "end_date",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date-time",
                            "description": "The end date for the performance period in ISO format.",
                            "title": "End Date"
                        },
                        "description": "The end date for the performance period in ISO format."
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/GeneralProjectPerformance"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Invalid credentials",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponseModel"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponseModel"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "ErrorResponseModel": {
                "properties": {
                    "detail": {
                        "type": "string",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "required": ["detail"],
                "title": "ErrorResponseModel",
                "description": "Standard error response model."
            },
            "GeneralProjectPerformance": {
                "properties": {
                    "project_id": {
                        "type": "integer",
                        "title": "Project Id"
                    },
                    "project_name": {
                        "type": "string",
                        "title": "Project Name"
                    },
                    "commits": {
                        "type": "integer",
                        "title": "Commits"
                    },
                    "additions": {
                        "type": "integer",
                        "title": "Additions"
                    },
                    "deletions": {
                        "type": "integer",
                        "title": "Deletions"
                    },
                    "changes": {
                        "type": "integer",
                        "title": "Changes"
                    },
                    "contributors": {
                        "items": {
                            "$ref": "#/components/schemas/ProjectContributorStats"
                        },
                        "type": "array",
                        "title": "Contributors"
                    },
                    "daily_commit_counts": {
                        "additionalProperties": {
                            "type": "integer"
                        },
                        "type": "object",
                        "title": "Daily Commit Counts"
                    },
                    "daily_additions": {
                        "additionalProperties": {
                            "type": "integer"
                        },
                        "type": "object",
                        "title": "Daily Additions"
                    },
                    "daily_deletions": {
                        "additionalProperties": {
                            "type": "integer"
                        },
                        "type": "object",
                        "title": "Daily Deletions"
                    },
                    "daily_changes": {
                        "additionalProperties": {
                            "type": "integer"
                        },
                        "type": "object",
                        "title": "Daily Changes"
                    }
                },
                "type": "object",
                "required": [
                    "project_id",
                    "project_name",
                    "commits",
                    "additions",
                    "deletions",
                    "changes",
                    "contributors",
                    "daily_commit_counts",
                    "daily_additions",
                    "daily_deletions",
                    "daily_changes"
                ],
                "title": "GeneralProjectPerformance",
                "description": "Top-level structure returned by the performance service for a project."
            },
...
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "MembersResponse": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "title": "Id",
                        "description": "User ID"
                    },
                    "username": {
                        "type": "string",
                        "title": "Username",
                        "description": "Username"
                    },
                    "name": {
                        "type": "string",
                        "title": "Name",
                        "description": "Full name"
                    },
                    "web_url": {
                        "type": "string",
                        "title": "Web Url",
                        "description": "Web URL to the user's profile"
                    },
                    "avatar_url": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Avatar Url",
                        "description": "URL to the user's avatar"
                    },
                    "state": {
                        "type": "string",
                        "title": "State",
                        "description": "Account state"
                    },
                    "access_level": {
                        "type": "integer",
                        "title": "Access Level",
                        "description": "Access level in the project"
                    },
                    "access_level_name": {
                        "type": "string",
                        "title": "Access Level Name",
                        "description": "Access level name"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "username",
                    "name",
                    "web_url",
                    "state",
                    "access_level",
                    "access_level_name"
                ],
                "title": "MembersResponse",
                "description": "Response schema for GitLab project members."
            },
...
            "ProjectContributorStats": {
                "properties": {
                    "author_name": {
                        "type": "string",
                        "title": "Author Name"
                    },
                    "author_email": {
                        "type": "string",
                        "title": "Author Email"
                    },
                    "commits": {
                        "type": "integer",
                        "title": "Commits"
                    },
                    "additions": {
                        "type": "integer",
                        "title": "Additions"
                    },
                    "deletions": {
                        "type": "integer",
                        "title": "Deletions"
                    },
                    "changes": {
                        "type": "integer",
                        "title": "Changes"
                    }
                },
                "type": "object",
                "required": [
                    "author_name",
                    "author_email",
                    "commits",
                    "additions",
                    "deletions",
                    "changes"
                ],
                "title": "ProjectContributorStats",
                "description": "Contributor-specific stats within a project. (authors of commits)"
            },
...
            "ProjectsResponse": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "title": "Id",
                        "description": "Project ID"
                    },
                    "name": {
                        "type": "string",
                        "title": "Name",
                        "description": "Project name"
                    },
                    "name_with_namespace": {
                        "type": "string",
                        "title": "Name With Namespace",
                        "description": "Project name with namespace"
                    },
                    "path_with_namespace": {
                        "type": "string",
                        "title": "Path With Namespace",
                        "description": "Project path with namespace"
                    },
                    "tag_list": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "title": "Tag List",
                        "description": "List of project tags"
                    },
                    "topics": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "title": "Topics",
                        "description": "List of project topics"
                    },
                    "web_url": {
                        "type": "string",
                        "title": "Web Url",
                        "description": "Web URL to the project"
                    },
                    "avatar_url": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Avatar Url",
                        "description": "URL to the project's avatar"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Created At",
                        "description": "Project creation timestamp"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "name",
                    "name_with_namespace",
                    "path_with_namespace",
                    "tag_list",
                    "topics",
                    "web_url",
                    "created_at"
                ],
                "title": "ProjectsResponse",
                "description": "Response schema for GitLab projects."
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": ["loc", "msg", "type"],
                "title": "ValidationError"
            },
...
        },
        "securitySchemes": {
            "HTTPBearer": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}